requirements.txt
discord
dnspython
pymongo
seaborn
matplotlib
Pillow
psutil
